requirements.txt
pandas==1.4.1
matplotlib==3.7.3
seaborn==0.12.2
Pillow==9.2.0
numpy==1.23.1
scipy==1.4.1
latex==0.7.0